Resume_analyzer/requirements.txt
PyPDF2
python-docx